requirements.txt
# Core FastAPI and WebSocket dependencies
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic

# Database dependencies
sqlalchemy
psycopg2-binary
alembic
asyncpg

# AI and Machine Learning libraries
google-genai
openai

# Data processing and utilities
python-dotenv
pydub

# Additional utility libraries
pytz
python-dateutil

# HTTP and API utilities
httpx
requests
aiofiles

# Audio and media processing
pillow
opencv-python

# Redis for caching (optional)
redis
aioredis

# Development and debugging
pytest
black
flake8

# Scheduler for daily tasks
apscheduler

# Additional dependencies found in the codebase
typing
datetime
pathlib
secrets
enum
contextlib 